#52 Remove Roxy from all namespaces
This is still a bit of a work in progress.  Everything seems to work
when testing and building using gradle from the command line.  However,
the ParameterizedTest class isn't detecting test classes from inside
IDEA.

While updating the namespaces, I moved the assertion functions out into
a separate assert library.  Since everyone's imports are already going
to break with the namespace change, I felt this was a good time to re-
organize the code so that we have better separation of internal and
external libraries.