Custom Failure Messages in Description, not Data
Moved failure message content into the description argument, of
fn:error() instead of relying on the data argument to fn:error().

Updated assert-throws-error-with-message() to assert against the
message element in the failure stack instead of format-string.  Added
a few new tests cases for this, and updated the existing tests.

Added custom failure message to assert-throws-error-with-message()