Merged marklogic-junit library into marklogic-unit-test
Bumped version to 0.13.develop. Once this is ready for 0.13.0 or 1.0.0, will need to update the references to 0.13.develop. 

Also made a fix in marklogic-unit-test-client/build.gradle so that createHttpCredentials isn't executed when any task is run, but only when it's run. 

See the README file in the marklogic-junit directory for more information.