Releasing 1.0.0 (#89)
* Issue #44: assert-equal-json inconsistent in comparisons (#85)

Fixing issue where empty sequences were not being considered equal to one another when using assert-equal-json().

* bintray config fixes

I am running "gradle -i bintrayUpload" from the root project directory, and all 8 artifacts are published seemingly successfully, but the bintray plugin still throws an error. The notes on the 1.8.4 release imply that 1.8.4 fixes a bug regarding publishing something multiple times, but that appears to be the error. But, the packages still seem to be uploaded successfully.

* Updating Home page
- Description of marklogic-unit-test and marklogic-junit
- Links to sample projects
- Moving contributing data to wiki

* Fixing cvsUrl for each project

* Renaming marklogic-junit to marklogic-junit5 (#78)

This allows for marklogic-junit5 to be published under marklogic-community

* Improving the project landing page (#73)

Adding instructions for getting started using marklogic-unit-test.

* Fixing Bug #79 with Suite Names in Test names (#83)

- Test names that started with the suite name no longer throw a module
  not found error.

* #79 Updating assertion to be more specific (#84)

- Adding a better failure message

* Releasing Version 1.0.0

* Fixing Travis CI error (#90)

Changing Travis CI distribution to Trusty which still supports Oracle JDK 8

* fix AbstractMarkLogicTest.java javadocs (#93)

Correcting the JavaDocs. Several methods had @param that did not exist

* Update CONTRIBUTING.md (#94)

Questions or Problem linking to non-existent anchor. Removing it.

* Apply $params in xdmp:xslt-invoke() for coverage report format transforms (#95)

Can now take advantage of the test-dir parameter in an XSLT transform.

* Change test from name() to use self axis (#96)

name() will return the namespace-prefix of the instance XML document, which may not use "error". Changing to use the self:: axis and use XPath to test if it is an error:error element.

* Update XmlNode.java to use the diamond operator <> (#97)

No need to repeat the type on the right hand side

* new ByteArrayInputStream(xml.getBytes()) relies upon the default character encoding of the system to convert the string to a byte array. (#98)

Instead, use new InputSource(new StringReader(xml))

The StringReader avoids having to convert to a byte array and will send the character stream to the InputSource, which will determine the encoding through standard XML API parse rules.

* Abstract mark logic test logger (#99)

* Remove unnecessary declaration of logger that is shadowing the logger from LoggingObject that it extends.

* Remove unused imports (now that logger is not explicitly initialized here)

* #46 Change test failure stack output from encoded XML to a tabbed text format (#61)

* Updating 1.0.beta references to 1.0.0 (#101)

* Updating MarkLogic Copyrights to 2019 (#102)

* Removing Roxy reference (#103)

* Updating Change Log for 1.0.0