Adding support for Maven Central and updating dependencies (#134)
Can now publish all 3 modules to Maven Central - though I'm not 100% sure about the modules project. publishToMavenLocal works fine, but won't know if Maven Central accepts it until we try it. It's a bit different of course because it contains modules and not Java code, and thus the javadoc jar is empty. 

Updated all the dependencies to latest versions, most notably Java Client to 5.5.0 and ml-javaclient-util to 4.3.0 (which is not yet published to Maven Central, but will be very soon - likely today). 

Also updated the marklogic-unit-test-client tests to use JUnit 5 instead of JUnit 4 . This has no impact on users though.

Co-authored-by: Rob Rudin <rob.rudin@marklogic.com>