DEVEXP-522 New docs outline and intro
May change the intro as more pages are written, but this should be good for now. 

Dropped a bunch of boilerplate from the CONTRIBUTING - this will be further improved soon to focus more on actually developing and testing this project. 

Removed most of the README as well, so it follows our pattern of providing a brief intro and then sending the user to the user guide.