DEVEXP-525 Adding docs on running tests
I realized while doing this that we'll need a separate page soon for the JUnit 5 support. We have docs on that in the form of a README and an example project, but it'd be better to fold that into this user guide. That's outside the scope of this PR. 

I tossed in the stuff for loading data too, as that ended up being trivial to write.