#180 Improving labels in UI
Also fixing a bug where setup/teardown modules were being counted in the number of tests.

Also removed duplication of logic from default.xqy, which is now using the "is" functions in test-controller.xqy. Added a few of these functions for syntactic sugar as well, just to make the code easier to read in default.xqy. 

We unfortunately don't have any tests for the default.xqy module, so this is being tested manually via ml-gradle's unit-test-example project.