MLE-24704 Using ml-javaclient-util snapshot
Using this so that ml-gradle can temporarily depend on a snapshot for its tests. It will still compile against 1.5.0 since that's a compileOnly dependency. But this needs to be released after ml-gradle so that it can depend on 6.1.0 of ml-javaclient-util.

Also took a crack at a proper NOTICE file, will run this by Legal soon.